+++ .gitignore
@@ -1,5 +1,6 @@
 .DS_Store
 .ipynb_checkpoints/
+__MACOSX/
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
